refactor(system): extract QueryOrchestrator from JarvisSystem
Move ask(), _detect_agent_intent(), _run_agent(), and _build_tools()
(~400 lines of behavior) into a standalone QueryOrchestrator class.
JarvisSystem.ask() and its helpers become thin delegates.

This is the key step toward isolated unit testing: QueryOrchestrator
can be constructed with a Mock(spec=JarvisSystem) without materializing
the full subsystem graph.

Update two tests in tests/operators/test_operators.py to patch
QueryOrchestrator._run_agent instead of JarvisSystem._run_agent.

Part 3 of 4 in the JarvisSystem decomposition (issue #226).
